Conflicts:
	ground/openpilotgcs/src/plugins/hitl/hitloptionspage.ui
	ground/openpilotgcs/src/plugins/hitl/simulator.cpp
	shared/uavobjectdefinition/groundtruth.xml

+++ ground/openpilotgcs/src/libs/opmapcontrol/src/internals/projections/mercatorprojectionyandex.h
@@ -40,7 +40,7 @@ public:
     virtual double Axis() const;
     virtual double Flattening()const;
     virtual core::Point FromLatLngToPixel(double lat, double lng, int const& zoom);
-    virtual internals::PointLatLng FromPixelToLatLng(const int &x,const int &y,const int &zoom);
+    virtual internals::PointLatLng FromPixelToLatLng(const qint64 &x,const qint64 &y,const int &zoom);
     virtual  Size GetTileMatrixMinXY(const int &zoom);
     virtual  Size GetTileMatrixMaxXY(const int &zoom);
 private:
